requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-chroma>=0.1.0
langchain-community>=0.2.0
langgraph
streamlit>=1.35.0
chromadb>=0.5.0
python-dotenv>=1.0.0
pydantic>=2.0.0
